requirements.txt
llama-index==0.14.5
python-dotenv==1.0.1
tree-sitter-languages==1.10.2
tree-sitter==0.23.2
llama-index-vector-stores-elasticsearch==0.5.1
llama-index-embeddings-openai==0.5.1
llama-index-llms-ollama==0.8.0
spacy==3.8.7
IPython==8.18.1
tabulate==0.9.0
streamlit==1.50.0
httpx>=0.25.0
# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
# OpenAI API for hybrid assistant
openai>=1.0.0
# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0